refactor: move booting env with fixtures in test-helpers utility module
